Untested: Make ID attributes for star pagination.
This should give ID attributes to page numbers in Columbia opinions, so that it's possible to link directly to a particular page by appending, e.g., "#page37" to the URL.

I haven't managed to get the test server running yet, so I don't know if this breaks anything, but I can't imagine that it would. I don't mean to create extra work for anyone, and I'm happy to try again and test it once I have more time, but I figured I'd submit it in the meantime since it's such a simple change.