Convert opinion page numbers into anchor links
This regex should replace:
'<page_number>37</page_number>'
with:
' <a class="star-pagination" id="page37" href="#page37">*37</a>'
This would make it so users can right-click star pagination and copy the link to that specific part of the page. We may or may not want to adjust the CSS as well, to make that feature more apparent.